Move implementation of /ruler/delete_tenant_config to Ruler.
